frontend/requirements.txt
langchain-gigachat
langgraph
psycopg2-binary
configparser
sqlalchemy
flask
gTTS
SpeechRecognition
pydub